Extract validator input helpers
